add scope SCOPE_ZAKEN_GEFORCEERD_BIJWERKEN
